pyproject.toml
[project]
name = "noopt"
version = "0.1.0"
description = "Option analysis with Schwab"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.2.6",
    "polars>=1.30.0",
    "pydantic>=2.11.6",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
authors = [
    { name = "Cameron Russell" }
]

[build-system]
requires = ["uv_build>=0.7.13,<0.8.0"]
build-backend = "uv_build"
